# perf: optimize segment tree serialization.