@abeaumont 2022-01-10
The problem becomes more clear when looking at the number of allocated objects: 
![segment-serialize-allocs](https://user-images.githubusercontent.com/80059/148801450-76d4a8a1-0054-4998-8f5c-b941f3883de7.png)