@abeaumont 2022-01-12
I deployed the branch to dev (thanks @kolesnikovae) to check the optimization in https://demo-dev.pyroscope.io.

This is how CPU consumption changes, with 1h ranges in both left and right:
![dev-demo-cpu](https://user-images.githubusercontent.com/80059/149182270-edaa28ec-8ceb-41a8-a217-d8509167144a.png)

Even the red runtime values are better in absolute terms.

For allocation objects, I used the comparison view which shows better how allocations mostly disappear now in segment serialization:

![dev-demo-alloc-objects](https://user-images.githubusercontent.com/80059/149182778-9c692d17-d704-46eb-95a1-5c51b9c22539.png)